feat: [api-handlers-decomposition-recovery-006] - Retire monolithic handlers.go and satisfy size gates
